"fix" warning on type conversion
The interface for setting the leds on and off I used uint8_t type, as
the number of leds is 16. However later on there is a test that is
intended to stress this aspect (out of bound) that uses int as input
led_number and so the compiler keeps warning about the conversion from
int to uint8_t.

To avoid this warning, this change replaces uint8_t with int in
led_driver interface.